requirements.txt
pyswisseph==2.08.00-1
